src/functions/worker/requirements.txt
azure-functions
azure-servicebus
azure-cosmos
